🎨 refactor : Update directory structure for improved readability
