--- app/Main/Domain/Restaurant.php
@@ -1,24 +1,0 @@
-<?php
-// 식당 도메인 모델
-namespace App\Main\Domain;
-
-use Illuminate\Database\Eloquent\Model;
-
-class Restaurant extends Model
-{
-  // 식당 도메인 모델의 정의
-  // 식당과 관련된 로직을 구현할 수 있습니다.
-
-  public static function searchByLocationAndGenre(string $location, string $genre)
-  {
-    // 지역과 음식 장르에 맞는 식당 검색 로직
-    return self::where('location', $location)->where('genre', $genre)->get();
-  }
-
-  // FIXME: 미구현
-  public static function findNearby()
-  {
-    // 주변 식당 검색 로직
-    return self::where('latitude', $latitude)->where('longitude', $longitude)->get();
-  }
-}
