+++ routes/api.php
@@ -2,8 +2,8 @@
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
-use App\Main\Actions\SearchRestaurantsAction;
-use App\Main\Actions\FindNearbyRestaurantsAction;
+use App\Search\Actions\SearchRestaurantsAction;
+use App\Search\Actions\FindNearbyRestaurantsAction;
 
 /*
 |--------------------------------------------------------------------------
@@ -17,14 +17,17 @@ use App\Main\Actions\FindNearbyRestaurantsAction;
 */
 
 /**
- * 가게 검색 기능(장르, 대형 지역, 중형 지역 선택 가능)
+ * 가게 검색 기능(장르, 대형 지역, 중형 지역, 가게명를 선택하여 검색 가능)
  */
 Route::get('/search', function (Request $request, SearchRestaurantsAction $action) {
+    // request에서 각 요소를 가져옴
     $genre = $request->input('genre');
     $large_area = $request->input('large_area');
     $middle_area = $request->input('middle_area');
+    $keyword = $request->input('name');
 
-    return $action($genre, $large_area, $middle_area);
+    // 검색 액션 실행하여 결과 반환
+    return $action($genre, $large_area, $middle_area, $keyword);
 });
 
 /**
@@ -37,21 +40,6 @@ Route::get('/search/nby', function (Request $request, FindNearbyRestaurantsActio
     $range = 5; // 기본 검색 범위 (5km)
 
     return $action($lat, $lng, $range);
-});
- 
-/**
- * 가게명으로 검색 기능
- */
-Route::get('/search/name', function (Request $request, \App\Services\RecruitApiService $recruitApiService) {
-    $name = $request->input('name');
-
-    if ($name) {
-        $response = $recruitApiService->searchRestaurantsByName($name);
-    } else {
-        $response = null;
-    }
-
-    return response()->json($response);
 });
 
 // TODO: 사용자 성향에 따른 가게 추천
